backend/requirements.txt
# SecureLicenseSystem Backend Dependencies
# Install with: pip install -r requirements.txt

flask>=3.0.0
flask-cors>=4.0.0
cryptography>=41.0.0
PyJWT>=2.8.0
